% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfp_predict_oos.R
\name{bsfp.predict.oos}
\alias{bsfp.predict.oos}
\title{Out-of-sample prediction for BSFP}
\usage{
bsfp.predict.oos(
  bsfp.fit,
  test_data,
  response_type = "continuous",
  model_params = NULL,
  nsample,
  progress = TRUE,
  starting_values = NULL
)
}
\arguments{
\item{bsfp.fit}{Results from fitting \code{bsfp} on training data.}

\item{test_data}{Matrix-list dataset of held-out test data.}

\item{response_type}{Continuous or binary response. Must be one of 'continuous' (deafult) or 'binary'.}

\item{model_params}{May be \code{NULL} if \code{model_params=NULL} in \code{bsfp} fit.
Otherwise, specify as \code{(error_vars, joint_vars, indiv_vars, beta_vars, response_vars)}.}

\item{nsample}{Integer specifing number of Gibbs sampling iterations}

\item{progress}{Boolean determining if progress of the sampler be displayed}

\item{starting_values}{List of starting values for \eqn{\mathbf{V}, \mathbf{U}_s, \mathbf{W}_s, \mathbf{V}_s} for \eqn{s=1,\dots, q}.
If \code{NULL}, initialize from prior.}
}
\value{
Returns a list with the following parameters:
\item{test_data}{Test data provided by user}
\item{EY.draw}{List of posterior samples for the E(Y|X), i.e. \eqn{\beta_0 + \mathbf{V}\boldsymbol{\beta}_{joint} + \sum_{s=1}^q \mathbf{V}_s \boldsymbol{\beta}_s} for each Gibbs sampling iteration.}
\item{V.draw}{List of posterior samples for joint scores, \eqn{\mathbf{V}}}
\item{U.train}{List of posterior samples for joint loadings for each source, \eqn{\mathbf{U}_s} for \eqn{s=1,\dots,q} given by the training BSFP fit}
\item{W.train}{List of posterior samples for individual loadings for each source,  \eqn{\mathbf{W}_s} for \eqn{s=1,\dots,q} given by the training BSFP fit}
\item{Vs.draw}{List of posterior samples for individual scores for each source, \eqn{\mathbf{V}_s} for \eqn{s=1,\dots,q}}
\item{ranks}{Vector with the estimated joint and individual ranks. \code{ranks[1]} is the estimated joint rank. \code{ranks[2:(q+1)]} correspond to the individual ranks for each source.}
\item{tau2.train}{List of posterior samples for the response variance if the response was continuous given by training BSFP fit}
\item{beta.train}{List of posterior samples for the regression coefficients used in the predictive model given by training BSFP fit}
\item{Xm.draw}{List of posterior samples for missing predictors imputations}
}
\description{
Modified version of the function \code{bsfp.predict} from the GitHub repo BSFP for out-of-sample predictions.
}
\details{
Generate new scores for held-out test data based on a
training fit of BSFP. Uses the estimated ranks and joint and individual loadings. Cannot
be used if missing values are present in test data.
}
