% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_preprocessing.R
\name{cdf_transform}
\alias{cdf_transform}
\title{apply cdf transform to multi-view predictors}
\usage{
cdf_transform(Z_m, Z_m_test = NULL, smoothed = F)
}
\arguments{
\item{Z_m}{Train set predictors}

\item{Z_m_test}{Test set predictors}

\item{smoothed}{Use smoothed cdf}
}
\value{
List of transformed features
}
\description{
apply cdf transform to multi-view predictors
}
