% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_predictions.R
\name{coeff_JAFAR}
\alias{coeff_JAFAR}
\title{Compute the induced regression coefficients in the linear predictor of y|X=x for all MCMC iteration}
\usage{
coeff_JAFAR(risMCMC, rescale_pred = FALSE)
}
\arguments{
\item{risMCMC}{Output of the Gibbs Sampler for JAFAR under the D-CUSP prior}

\item{rescale_pred}{Rescale loadings to induce correlation matrix}
}
\value{
A list with the values of regression coefficients for each view in the linear predictor of y|X=x for all MCMC iteration
}
\description{
Compute the induced regression coefficients in the linear predictor of y|X=x for all MCMC iteration
}
