% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfp_predict_oos.R
\name{data.rearrange}
\alias{data.rearrange}
\title{Original function \code{data.rearrange} from the GitHub repo BSFP}
\usage{
data.rearrange(data, rmt = F, sigma = NULL)
}
\description{
Original function \code{data.rearrange} from the GitHub repo BSFP
}
\author{
Jun Young Park (2020)
}
