% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_preprocessing.R
\name{F_hat}
\alias{F_hat}
\title{Point-wise evaluation empirical cdf function}
\usage{
F_hat(xNew, xTrain)
}
\arguments{
\item{xNew}{location at which the cdf function is to be evaluated}

\item{xTrain}{data to construct the empirical cdf}
}
\description{
Point-wise evaluation empirical cdf function
}
