% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jafar_preprocessing.R
\name{features_reorder_HC}
\alias{features_reorder_HC}
\title{predictors preprocess: reorder features via hierarchical clustering for better visualization}
\usage{
features_reorder_HC(X_m, X_m_test = NULL, K0_HC = 15)
}
\arguments{
\item{X_m}{Train set predictors}

\item{X_m_test}{Test set predictors}

\item{K0_HC}{Reference number of clusters for hierarchical clustering (default: 15)}
}
\value{
List of preprocessed features and rescaling factors
}
\description{
predictors preprocess: reorder features via hierarchical clustering for better visualization
}
