% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_predictions.R
\name{get_coeff_JAFAR}
\alias{get_coeff_JAFAR}
\title{Compute the induced regression coefficients in the linear predictor of y|X=x for a single MCMC iteration}
\usage{
get_coeff_JAFAR(
  M,
  K,
  K_m,
  p_m,
  Theta,
  s2_inv_y,
  Lambda_m,
  Gamma_m,
  s2_inv_m,
  rescale_pred = FALSE
)
}
\arguments{
\item{M}{number of views}

\item{K}{Number of latent factors in view-specific components (length M)}

\item{p_m}{views dimensions (length M)}

\item{Theta}{Response loadings s2_inv_y}

\item{s2_inv_y}{Response noise precision}

\item{Lambda_m}{Shared-component loadings}

\item{Gamma_m}{View-specific loadings}

\item{s2_inv_m}{Precisions of idiosyncratic components in multi-view features}

\item{rescale_pred}{Rescale loadings to induce correlation matrix}
}
\value{
A list with the values of regression coefficients for each view in the linear predictor of y|X=x for a single MCMC iteration
}
\description{
Compute the induced regression coefficients in the linear predictor of y|X=x for a single MCMC iteration
}
