% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_preprocessing.R
\name{get_F_smooth}
\alias{get_F_smooth}
\title{Smoothed version of the empirical cdf}
\usage{
get_F_smooth(vec)
}
\arguments{
\item{vec}{Predictor values}
}
\description{
Smoothed version of the empirical cdf
}
