% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_preprocessing.R
\name{get_NA_X}
\alias{get_NA_X}
\title{Identification of missing entries in multi-view predictors}
\usage{
get_NA_X(X_m)
}
\arguments{
\item{X_m}{Multi-view predictors}
}
\value{
View-wise list if indices of meassing features
}
\description{
Identification of missing entries in multi-view predictors
}
