% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_preprocessing.R
\name{get_Q_smooth}
\alias{get_Q_smooth}
\title{Smoothed version of the quantile cdf}
\usage{
get_Q_smooth(vec)
}
\arguments{
\item{vec}{Predictor values}
}
\description{
Smoothed version of the quantile cdf
}
