% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_gibbs_sampler.R
\name{gibbs_JAFAR_CUSP}
\alias{gibbs_JAFAR_CUSP}
\title{Run Gibbs Sampler for JAFAR under the D-CUSP prior}
\usage{
gibbs_JAFAR_CUSP(
  y,
  X_m,
  n,
  M,
  p_m,
  binary_y = FALSE,
  seed = 123,
  nMCMC = 2500,
  nBurnIn = 5000,
  nThin = 10,
  Kmax = NULL,
  Kmax_m = NULL,
  K0 = NULL,
  K0_m = NULL,
  prec0 = NULL,
  prec0m = NULL,
  a_m = NULL,
  b_m = NULL,
  a_sig = NULL,
  b_sig = NULL,
  a_theta = NULL,
  b_theta = NULL,
  a_chi = NULL,
  b_chi = NULL,
  var_spike_theta = NULL,
  a_xi = NULL,
  b_xi = NULL,
  var_spike = NULL,
  alpha = NULL,
  alpha_loc = NULL,
  t0 = -1,
  t1 = -5e-04,
  t0_adapt = 20,
  get_latent_vars = TRUE,
  rescale_pred = FALSE,
  get_last_sample = FALSE
)
}
\arguments{
\item{y}{vector of responses (length n)}

\item{X_m}{list of M multi-view predictors (the m-th elment if a matrix of dim. n x p_m\link{m})}

\item{n}{number of observations}

\item{M}{number of views}

\item{p_m}{views dimensions (length M)}

\item{binary_y}{Is the response binary?}

\item{seed}{See for random number generation}

\item{nMCMC}{Number of total MCMC iterations}

\item{nBurnIn}{Number of burn-in iterations}

\item{nThin}{Thinning of MCMC iterations}

\item{Kmax}{Maximum number of latent factors in shared components}

\item{Kmax_m}{Maximum number of latent factors in view-specific components (length M)}

\item{K0}{Initial number of latent factors in shared components}

\item{K0_m}{Initial number of latent factors in view-specific components (length M)}

\item{prec0}{Prior precision for response intercept}

\item{prec0m}{Prior precisions for predictors intercepts}

\item{a_m}{Shape parameters of inverse-gamma prior on predictors idiosyncratic components}

\item{b_m}{Scale parameters of inverse-gamma prior on predictors idiosyncratic components}

\item{a_sig}{Shape parameter of inverse-gamma prior on response noise}

\item{b_sig}{Scale parameter of inverse-gamma prior on response noise}

\item{a_theta}{Shape parameter in slab element of prior for response loadings}

\item{b_theta}{Scale parameter in slab element of prior for response loadings}

\item{a_chi}{Shape parameters in slab element of prior fo shared-component loadings}

\item{b_chi}{Scale parameters in slab element of prior for shared-component loadings}

\item{var_spike_theta}{Variance parameter in spike element of prior for response-loadings}

\item{a_xi}{Shape1 parameters in beta of prior on response-loadings spike and slab weights}

\item{b_xi}{Shape2 parameters in beta of prior on response-loadings spike and slab weights}

\item{var_spike}{Variance parameter in spike element of prior for predictors-loadings}

\item{alpha}{Stick-breaking parameter in shared-component loadings}

\item{alpha_loc}{Stick-breaking parameter in view-specific loadings}

\item{t0}{Intercept in MCMC adaptation}

\item{t1}{Slope in MCMC adaptation}

\item{t0_adapt}{Off-set of MCMC adaptation}

\item{get_latent_vars}{Include in output of latent variables form factor model}

\item{rescale_pred}{Rescale loadings to induce correlation matrix in inputing missing values}

\item{get_last_sample}{Include in output the last sample}
}
\value{
A list of output from the MCMC chain
}
\description{
Run Gibbs Sampler for JAFAR under the D-CUSP prior
}
