% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfr_gibbs.R
\name{gibbs_jfr}
\alias{gibbs_jfr}
\title{Gibbs sampler for \code{jfr}}
\usage{
gibbs_jfr(
  X_m,
  y = NULL,
  yBinary = F,
  K0 = NULL,
  tMCMC = 20000,
  tBurnIn = 15000,
  tThin = 10,
  hyperparams = list(),
  get_latent_vars = TRUE,
  get_last_sample = FALSE,
  parallel = TRUE,
  tempered = FALSE,
  rescale_pred = FALSE
)
}
\arguments{
\item{X_m}{Multi-view input data, pre-processed via \code{\link{preprocess_X}}.
List of length \code{M}; m-th element: matrix \code{n x p_m[m]}.
Rows should correspond to samples, columns to features.}

\item{y}{Vector of responses (length \code{n}) pre-processed via \code{\link{preprocess_y}}.
Set to \code{NULL} for unsupervised mode (default: \code{NULL}).}

\item{yBinary}{Logical, indicating if the response(s) are binary (default: \code{FALSE}).}

\item{K0}{Upper bound to numbers of latent factors (optional)
If \code{NULL}, \code{K0} is set to \code{3*log(max(p_m))}}

\item{tMCMC}{Total number of MCMC iterations (default: 20000).}

\item{tBurnIn}{Number of burn-in iterations (default: 15000).}

\item{tThin}{Thinning interval for saving samples (default: 10).}

\item{hyperparams}{List of hyperparameters for the \code{I-CUSP} prior distributions.
Missing hyperparameters are replaced by default values encoded in \code{\link{set_hyperparameters}}.}

\item{get_latent_vars}{Return latent factors and loading matrices (logical, default: \code{TRUE}).}

\item{get_last_sample}{Return the last sample of the MCMC chain (logical, default: \code{FALSE}).}

\item{parallel}{Use parallel computation for the loadings update (logical, default: \code{TRUE}).}

\item{tempered}{Use tempered full-conditional for the loadings matrices (logical, default: \code{FALSE}).}

\item{rescale_pred}{Rescale loadings when computing response predictions (logical, default: \code{FALSE}).}
}
\value{
A list containing posterior samples, latent variables (if requested), and other relevant model outputs.
}
\description{
Fits a Bayesian Joint Factor Regression (\code{jfr}) model using Gibbs sampling.
Variation across multiple data views is explained by a single set of global latent factors.
The model can be fitted in both unsupervised and supervised settings.
Default and optional outputs include posterior means of the induced covariances,
posterior samples of residual variances, latent factors, and factor loadings.
Supports parallel computation and tempered loading updates to limit rank estimation in extreme large-p-small-n settings.
}
\details{
\itemize{
\item Ensure that all matrices in \code{X_m} have the same number of rows (subjects).
\item Missing data in \code{X_m} are allowed as \code{NA} and imputed in the MCMC.
}
}
\note{
All posterior samples are reported only after burn-in, except for \code{K}.
The number of samples after thinning is \code{tFull=tMCMC\%/\%tThin} and \code{tEff=(tMCMC-tBurnIn)\%/\%tThin}
for the full chain and post burn-in, respectively. \cr
The output list includes:
\itemize{
\item{\code{K}: Number shared latent factors (vector of length \code{tFull}).}
\item{\code{K_Lm_eff}: Numbers of latent factors active in each view (matrix \code{tEff x M}).}
\item{\code{active_Lm}: Binary indicators of latent factors activity across views (binary array \code{tEff x K x M}).}
\item{\code{Cov_m_mean}: Posterior mean of the covariance matrix for each dataset (list of length \code{M}; m-th element: matrix \code{p_m[m] x p_m[m]}).}
\item{\code{Marg_Var_m}: Marginal variances of features (list of length \code{M}; m-th element: matrix \code{tEff x p_m[m]}).}
\item{\code{s2_inv_m}: Inverse residual variances across views (list of length \code{M}; m-th element: matrix \code{tEff x p_m[m]}).}
\item{\code{mu_m}: Features intercepts across views (list of length \code{M}; m-th element: matrix \code{tEff x p_m[m]}).}
\item{\code{hyper_param}: List of hyperparameters used for the model, including user-specified values and defaults ones were missing.}
}
If \code{is_supervised = TRUE}:
\itemize{
\item{\code{K_T_eff}: Numbers of latent factors active in the response (vector of length \code{tEff}).}
\item{\code{active_T}: Binary indicators of latent factors activity in the response (binary matrix \code{tEff x K}).}
\item{\code{s2_inv}: Response inverse residual variances (vector of length \code{tEff}).}
\item{\code{mu_y}: Response intercept (vector of length \code{tEff}).}
\item{\code{Theta}: Response loadings on latent factors (matrix \code{tEff x K}).}
\item{\code{y_MC}: Latent probit utilities (matrix \code{tEff x n}). (only if \code{yBinary = TRUE}).}
}
If \code{get_latent_vars = TRUE}:
\itemize{
\item{\code{Lambda_m}: Loadings matrices on latent factors (list of length \code{M}; m-th element: array \code{tEff x p_m[m] x K}).}
\item{\code{eta}: Latent factors (array \code{tEff x n x K}).}
}
If the input matrices \code{X_m} contain missing values:
\itemize{
\item{\code{Xm_MC}: Posterior samples of imputed values for missing entries.
A list of length \code{M}; the m-th element is itself a list (one per feature with missingness),
each containing an \code{tEff × n_miss} matrix of imputed values across MCMC iterations.}
\item{\code{na_idx}: List of length \code{M}; the m-th element gives the column indices of missing entries in \code{X_m[[m]]}.}
\item{\code{na_row_idx}: List of length \code{M}; the m-th element gives the corresponding row indices of missing entries in \code{X_m[[m]]}.}
}
If \code{get_last_sample = TRUE}:
\itemize{
\item{\code{last_sample}: List of posterior values of all parameters at the last MCMC iteration, including latent factors, loadings, residual variances, and hyperparameters.}
}
}
\references{
Anceschi N., Ferrari F., Dunson D. B., & Mallick H. (2025).
\emph{Bayesian Joint Additive Factor Models for Multiview Learning}.
\url{https://arxiv.org/abs/2406.00778}

Legramanti S., Durante D., & Dunson D. B. (2020).
\emph{Bayesian cumulative shrinkage for infinite factorizations}.
Biometrika, 107(3), 745-752. \url{https://doi.org/10.1093/biomet/asaa008}
}
