% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_preprocessing.R
\name{is.scalar}
\alias{is.scalar}
\title{Check in any given variable is a scalar}
\usage{
is.scalar(x)
}
\arguments{
\item{x}{input variable}
}
\value{
Boolean value
}
\description{
Check in any given variable is a scalar
}
