% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jafar_predictions.R
\name{jafar_coeff_y}
\alias{jafar_coeff_y}
\title{Response predictions and induced regression coefficients for JAFAR and JFR}
\usage{
jafar_coeff_y(Xpred, risMCMC, rescale_pred = FALSE)
}
\arguments{
\item{Xpred}{A list of \code{M} features' matrices, the m-th of dimension \code{nPred x p_m[m]} or possibly with missing (\code{X_m[[m]]=NULL)}).}

\item{risMCMC}{Output of \code{\link{gibbs_jafar}} or \code{\link{gibbs_jfr}} containing posterior samples.}

\item{rescale_pred}{Logical, whether to rescale loadings when computing response predictions (default: \code{FALSE}).}
}
\value{
A list containing posterior samples of the predicted responses (matrix \code{tEff x nPred}),
and of the induced regression coefficients for each view (list of length \code{M}; m-th element: \code{tEff x p_m[m]}).
}
\description{
Response predictions and induced regression coefficients for JAFAR and JFR
}
