% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_preprocessing.R
\name{jafar_preprocess_X}
\alias{jafar_preprocess_X}
\title{JAFAR multi-view predictors preprocess: cdf transform + center & rescale}
\usage{
jafar_preprocess_X(Z_m, Z_m_test = NULL, copula = F, smoothed = F)
}
\arguments{
\item{Z_m}{Train set predictors}

\item{Z_m_test}{Test set predictors}

\item{copula}{Apply cdf transformation}

\item{smoothed}{Use smoothed cdf}
}
\value{
List of preprocessed features and rescaling factors
}
\description{
JAFAR multi-view predictors preprocess: cdf transform + center & rescale
}
