% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_preprocessing.R
\name{jafar_preprocess_y}
\alias{jafar_preprocess_y}
\title{JAFAR response preprocess: center & rescale}
\usage{
jafar_preprocess_y(yTrain, yTest = NULL)
}
\arguments{
\item{yTrain}{Train set responses}

\item{yTest}{Test set responses}
}
\value{
List of preprocessed responses and rescaling factors
}
\description{
JAFAR response preprocess: center & rescale
}
