% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jafar_initialization.R
\name{jafar_set_hyperparameters}
\alias{jafar_set_hyperparameters}
\title{Set Hyperparameters for JAFAR and JFR models}
\usage{
jafar_set_hyperparameters(hyperparams_list, M, is_supervised = FALSE)
}
\arguments{
\item{hyperparams_list}{A named list of hyperparameters to be used in the model.}

\item{M}{Integer, number of data-views.}

\item{is_supervised}{Logical, whether the model is supervised (default: FALSE).}
}
\value{
A named list of hyperparameters with defaults filled in where missing.
Scalar values are replicated \code{M} times where necessary.
}
\description{
Helper function to set hyperparameters for \code{\link{gibbs_jfr}} and \code{\link{gibbs_jafar}}.
Missing hyperparameters are assigned default values.
Supports both unsupervised and supervised (response-guided) settings.
}
\details{
Default hyperparameters include:
\itemize{
\item{\code{seed}: random seed for reproducibility (default: 123).}
\item{\code{t0, t1, t0_adapt}: adaptation parameters for MCMC (default: \code{t0=-1, t1=-5e-4, t0_adapt=200}).}
\item{\code{a_m, b_m}: shape and rate of inverse-gamma prior for idiosyncratic noise in each view.}
Scalars of vectors of length \code{M} (default: \code{a_m[m]=3, b_m[m]=1}).
\item{\code{prec0m}: precision of normal prior on intercepts.}
Scalar of vector of length \code{M} (default: \code{prec0m[m]=2}).
\item{\code{var_spike}: variance of normal spike in cusps.}
Scalar of vector of length \code{M} (default: \code{var_spike[m]=0.005}).
\item{\code{a_chi, b_chi}: hyperparameters for slab inverse-gamma prior in cusps.
Scalars of vectors of length \code{M} (default: \code{a_chi[m]=0.5, b_chi[m]=0.1}).}
\item{\code{alpha_L, alpha_G}: DP concentration parameters giving the expected number of factors, shared and local.
Scalars of vectors of length \code{M} (default: \code{alpha_L[m]=5, alpha_G[m]=5}).}
}
If \code{is_supervised = TRUE}, additional hyperparameters for the response model are
\itemize{
\item{\code{a_sig, b_sig}: shape and rate of inverse-gamma prior for idiosyncratic noise (default: \code{a_sig=3, b_sig=1}).}
\item{\code{prec0}: precision of normal prior on intercept (default: \code{prec0=2}).}
\item{\code{var_spike_y}: variance of normal spike (default: \code{var_spike_y=0.005}).}
\item{\code{a_theta, b_theta}: hyperparameters for slab inverse-gamma prior in the slab (default: \code{a_theta=0.5, b_theta=0.1}).}
\item{\code{a_xi, b_xi}: shape parameters for beta prior on mixture weight in response loadings (default: \code{a_xi=3, b_xi=2}).}
}
}
