% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_postprocessing.R
\name{multivew_varimax}
\alias{multivew_varimax}
\title{Multi-view varimax routine}
\usage{
multivew_varimax(L, p_views, normalize = FALSE, eps = 1e-05, maxIter = 1000)
}
\arguments{
\item{p_views}{Number of features per view}

\item{normalize}{Normalization of loading columns (varimax internal parameter)}

\item{eps}{Threshold on trace}

\item{maxIter}{Maximum number of iterations}

\item{lambda}{One single sample of the concatenate shared-loadings}
}
\description{
Multi-view varimax routine
}
