% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_postprocessing.R
\name{multiview_MatchAlign}
\alias{multiview_MatchAlign}
\title{Multi-view MatchAlign routine}
\usage{
multiview_MatchAlign(lambda, eta, theta, p_views, normalize = F)
}
\arguments{
\item{lambda}{MCMC samples of the concatenate shared-loadings}

\item{theta}{MCMC samples of the shared latent factors}

\item{p_views}{Number of features per view}

\item{normalize}{Normalization of loading columns (varimax internal parameter)}

\item{risMCMC}{MCMC samples of the response loadings}
}
\description{
Multi-view MatchAlign routine
}
