% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_preprocessing.R
\name{order_index_na}
\alias{order_index_na}
\title{Empirical cdf transform with missing data}
\usage{
order_index_na(x)
}
\arguments{
\item{x}{Predictor values}
}
\description{
Empirical cdf transform with missing data
}
