% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_plots.R
\name{plot_coeff_jafar}
\alias{plot_coeff_jafar}
\title{Plot the induced regression coefficients in the linear predictor of y|X=x}
\usage{
plot_coeff_jafar(coef_jafar, out_folder = "~/Desktop")
}
\arguments{
\item{coef_jafar}{samples of the induced regression coefficients (output of \code{coeff_JAFAR})}

\item{out_folder}{Directory where to save output}
}
\description{
Plot the induced regression coefficients in the linear predictor of y|X=x
}
