% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_plots.R
\name{plot_cor_jafar}
\alias{plot_cor_jafar}
\title{Plot the empirical and inferred within-viewa correlation matrices}
\usage{
plot_cor_jafar(risMCMC, Z_m, out_folder = "~/Desktop/")
}
\arguments{
\item{risMCMC}{Output of the Gibbs Sampler for JAFAR under the D-CUSP prior}

\item{Z_m}{Train set multi-view predictors}

\item{out_folder}{Directory where to save output}
}
\description{
Plot the empirical and inferred within-viewa correlation matrices
}
