% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jafar_plots.R
\name{plot_correlations}
\alias{plot_correlations}
\title{Visualization of induced correlation matrices}
\usage{
plot_correlations(
  risMCMC,
  X_m = NULL,
  out_path = "~/Desktop/",
  out_name = "correlations"
)
}
\arguments{
\item{risMCMC}{Posterior samples, output of \code{\link{gibbs_jafar}} or \code{\link{gibbs_jfr}}}

\item{X_m}{Training set multi-view predictors (optional, default: \code{NULL}).
If \code{NULL}, only inferred correlation matrices are visualized.
If not \code{NULL}, the empirical correlation matrices are displayed besides the inferred ones}

\item{out_path}{Output path where the generated plot will be saved (default: "~/Desktop/")}

\item{out_name}{Output file name (default: "correlations")}
}
\description{
Plot the empirical and inferred within-view correlation matrices.
The induced correlations on \code{X_m} are obtained by marginalizing out all latent factors.
}
