% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jafar_plots.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Plot posterior means of factor loadings.}
\usage{
plot_loadings(
  risMCMC,
  out_path = "~/Desktop/",
  out_name_shared = "shared_loadings",
  out_name_specific = "specific_loadings"
)
}
\arguments{
\item{risMCMC}{Posterior samples, output of \code{gibbs_jafar} or \code{gibbs_jfr}}

\item{out_path}{Output path where the generated plot will be saved}

\item{out_name_shared}{Output file name for the shared component plot (default: "n_factors_shared")}

\item{out_name_specific}{Output file name for the specific components plot (default: "n_factor_specific")}
}
\description{
Rotational alignment must be performed in advanced through the function \code{multiviewMatchAlign}
}
