% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_plots.R
\name{plot_loadings_jafar}
\alias{plot_loadings_jafar}
\title{Plot the postprocessed shared-component loadings}
\usage{
plot_loadings_jafar(ris_PostProc, out_folder = "~/Desktop/")
}
\arguments{
\item{ris_PostProc}{Postprocessed shared component latent variables (output of \code{postprocess_JAFAR})}

\item{out_folder}{Directory where to save output}
}
\description{
Plot the postprocessed shared-component loadings
}
