% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_plots.R
\name{plot_n_fact_shared}
\alias{plot_n_fact_shared}
\title{Plot the of view-wise number of active shared factors throught the MCMC chain evolution}
\usage{
plot_n_fact_shared(risMCMC, out_folder = "~/Desktop")
}
\arguments{
\item{risMCMC}{Output of the Gibbs Sampler for JAFAR under the D-CUSP prior}

\item{out_folder}{Directory where to save output}
}
\description{
Plot the of view-wise number of active shared factors throught the MCMC chain evolution
}
