% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_plots.R
\name{plot_pred_jafar}
\alias{plot_pred_jafar}
\title{Plot out-of-sample predictions of the response's linear predictors}
\usage{
plot_pred_jafar(
  vPred,
  yTrue,
  s2y_inv = NULL,
  is_binary = F,
  out_folder = "~/Desktop/",
  size = 0.7,
  linewidth = 0.5,
  width = 0.7,
  alpha = 0.4,
  range = 0.25,
  filename = NULL
)
}
\arguments{
\item{vPred}{Predicted values}

\item{yTrue}{True values}

\item{s2y_inv}{MCMC samples of the response noise precision}

\item{is_binary}{Is the response binary?}

\item{out_folder}{Directory where to save output}

\item{size}{Point size}

\item{linewidth}{Linewidth of credible intervals bars}

\item{width}{Width of credible intervals extrema}

\item{alpha}{Opacity of credible intervals bars}

\item{range}{Jitter's range for binary response}

\item{filename}{Custom file name (optional)}
}
\description{
Plot out-of-sample predictions of the response's linear predictors
}
