% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_postprocessing.R
\name{postprocess_JAFAR}
\alias{postprocess_JAFAR}
\title{Postprocessing of shared-component latent variable via multi-view MatchAlign}
\usage{
postprocess_JAFAR(risMCMC, normalize_col = F, nBurnIn = 0)
}
\arguments{
\item{risMCMC}{Output of the Gibbs Sampler for JAFAR under the D-CUSP prior}

\item{normalize_col}{Normalization of loading columns (varimax internal parameter)}

\item{nBurnIn}{Number of (extra) MCMC iterations to skip}
}
\description{
Postprocessing of shared-component latent variable via multi-view MatchAlign
}
