% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jafar_predictions.R
\name{predict_y}
\alias{predict_y}
\title{Response predictions for \code{jafar} and \code{jfr}}
\usage{
predict_y(Xpred, risMCMC, rescale_pred = FALSE)
}
\arguments{
\item{Xpred}{A list of \code{M} features matrices, the m-th one of dimension \code{nPred x p_m[m]} or possibly missing (i.e. \code{X_m[[m]]=NULL}).}

\item{risMCMC}{Output of \code{\link{gibbs_jafar}} or \code{\link{gibbs_jfr}} containing posterior samples.}

\item{rescale_pred}{Rescale loadings when computing response predictions (logical, default: \code{FALSE}).}
}
\value{
A list containing posterior samples of the predicted responses (matrix \code{tEff x nPred}),
and of the induced regression coefficients for each view (list of length \code{M}; m-th element: \code{tEff x p_m[m]}).
}
\description{
Compute induced regression coefficients and predicted responses either in-sample or out-of-sample.
}
