% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jafar_predictions.R
\name{predict_y_raw}
\alias{predict_y_raw}
\title{Response predictions for JAFAR and JFR}
\usage{
predict_y_raw(Xpred, risMCMC, rescale_pred = FALSE)
}
\arguments{
\item{Xpred}{A list of \code{M} features' matrices, the m-th of dimension \code{nPred x p_m[m]} or possibly with missing (\code{X_m[[m]]=NULL)}).}

\item{risMCMC}{Output of \code{\link{gibbs_jafar}} or \code{\link{gibbs_jfr}} containing posterior samples.}

\item{rescale_pred}{Logical, whether to rescale loadings when computing response predictions (default: \code{FALSE}).}
}
\value{
A list containing posterior samples of the predicted responses (matrix \code{tEff x nPred}).
}
\description{
Response predictions for JAFAR and JFR
}
