% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jafar_preprocessing.R
\name{preprocess_X}
\alias{preprocess_X}
\title{predictors preprocess: center & rescale + cdf transform (optional)}
\usage{
preprocess_X(X_m, X_m_test = NULL, copula = F)
}
\arguments{
\item{X_m}{Train set predictors}

\item{X_m_test}{Test set predictors}

\item{copula}{Apply cdf transformation}
}
\value{
List of preprocessed features and rescaling factors
}
\description{
predictors preprocess: center & rescale + cdf transform (optional)
}
