% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_preprocessing.R
\name{Q_hat}
\alias{Q_hat}
\title{Point-wise evaluation empirical quantile function}
\usage{
Q_hat(pNew, xTrain)
}
\arguments{
\item{pNew}{probability at which the quantile function is to be evaluated}

\item{xTrain}{data to construct the empirical cdf}
}
\description{
Point-wise evaluation empirical quantile function
}
