% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jafar_preprocessing.R
\name{reorder_features}
\alias{reorder_features}
\title{Optional pre-process of the multi-view data}
\usage{
reorder_features(X_m, X_m_test = NULL, K0_HC = 15)
}
\arguments{
\item{X_m}{Multi-view data for the training set.
List of length \code{M}; m-th element: matrix \code{n x p_m[m]}.}

\item{X_m_test}{Multi-view data for the test set.
List of length \code{M}; m-th element: matrix \code{nTest x p_m[m]}.}

\item{K0_HC}{Reference number of clusters for hierarchical clustering (default: 15)}
}
\value{
List of pre-processed features and rescaling transformations
}
\description{
Reorder features via hierarchical clustering for better visualization.
Multi-view data for out-of-sample observations can optionally be provided as input.
If so, the corresponding features are re-ordered coherently with the training set.
}
\note{
The output list contains:
\itemize{
\item{\code{idx_sort_HC_m}: List of features reordering obtained by running hierarchical
clustering on empirical correlations.
This object can be used to restore the data to its original order.}
\item{\code{X_m}: Reordered multi-view features for the training set.}
\item{\code{X_m_test}: Reordered multi-view features for the test set (if provided).}
}
}
