% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_plots.R
\name{unif_jitter}
\alias{unif_jitter}
\title{Deterministic jitter for plotting out-of-sample binary responses}
\usage{
unif_jitter(vPred, range = 0.25)
}
\arguments{
\item{vPred}{Predicted values}

\item{range}{Jitter's range}
}
\description{
Deterministic jitter for plotting out-of-sample binary responses
}
