% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_updates.R
\name{update_loadings}
\alias{update_loadings}
\title{Update of view-wise loadings within the Gibbs sampler for JAFAR under the D-CUSP prior}
\usage{
update_loadings(
  n,
  p_m,
  K,
  K_Gm,
  X_m,
  facTfac,
  eta,
  phi_m,
  mu_m,
  s2_inv_m,
  chi_m,
  tau_m
)
}
\arguments{
\item{n}{Number of observations (integer)}

\item{p_m}{Dimension of the considered view (integer)}

\item{K}{Number of shared factors}

\item{K_Gm}{Number of view-specific factors}

\item{X_m}{Observed predictors in the considered view (real matrix: n x p_m)}

\item{eta}{Shared latent factors (real matrix: n x K)}

\item{phi_m}{View-specific latent factors (real matrix: n x K_Gm)}

\item{chi_m}{Shared-loadings prior variances (real positive vector: K)}

\item{tau_m}{View-specific loadings prior variances (real positive vector :K_Gm)}

\item{facTfac:}{t-crossp-roduct of overall factors (real matrix: (K+K_Gm) x (K+K_Gm))}

\item{mu_m:}{Intercepts (real vector: p_m)}

\item{s2_inv_m:}{Idiosyncrstic components precision (real positive vector: p_m)}
}
\value{
Updated loadings
}
\description{
Update of view-wise loadings within the Gibbs sampler for JAFAR under the D-CUSP prior
}
