% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAFAR_D-CUSP_predictions.R
\name{y_pred_JAFAR}
\alias{y_pred_JAFAR}
\title{Compute the linear predictor of y|X=x in out-of-sample observations for all MCMC iteration}
\usage{
y_pred_JAFAR(Xpred, risMCMC, rescale_pred = FALSE)
}
\arguments{
\item{Xpred}{Multi-view features in out-of-sample observations}

\item{risMCMC}{Output of the Gibbs Sampler for JAFAR under the D-CUSP prior}

\item{rescale_pred}{Rescale loadings to induce correlation matrix}
}
\value{
Linear predictors of y|X=x in out-of-sample observations for all MCMC iteration
}
\description{
Compute the linear predictor of y|X=x in out-of-sample observations for all MCMC iteration
}
