% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jafar_postprocessing.R
\name{multiviewMatchAlign}
\alias{multiviewMatchAlign}
\title{Perform rotational alignment using multi-view \code{MatchAlign}.}
\usage{
multiviewMatchAlign(ris_MCMC)
}
\arguments{
\item{risMCMC}{Posterior samples, as returned by \code{gibbs_jafar} or \code{gibbs_jfr}.}
}
\value{
A modified version of the input \code{risMCMC}, with latent factors, loading matrices,
and (if supervised) response loadings rotated according to multi-view \code{MatchAlign}.
}
\description{
Perform rotational alignment using multi-view \code{MatchAlign}.
}
