% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jafar_plots.R
\name{plot_coefficients}
\alias{plot_coefficients}
\title{Plot induced regression coefficients for y|X=x}
\usage{
plot_coefficients(yPred, out_path = "~/Desktop/", out_name = "coefficients")
}
\arguments{
\item{yPred}{Response predictions, output of \code{predict_y} or \code{predict_y_raw}}

\item{out_path}{Output path where the generated plot will be saved (default: "~/Desktop/")}

\item{out_name}{Output file name (default: "coefficients")}
}
\description{
Plot induced regression coefficients for y|X=x
}
