% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jafar_plots.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Visualization of loadings matrices}
\usage{
plot_loadings(
  risMCMC,
  out_path = "~/Desktop/",
  out_shared = "shared_loadings",
  out_specific = "specific_loadings"
)
}
\arguments{
\item{risMCMC}{Postprocesed posterior samples, output of \code{\link{multiviewMatchAlign}}}

\item{out_path}{Output path where the generated plot will be saved}

\item{out_shared}{File name for the shared component plot (default: "n_factors_shared")}

\item{out_specific}{File name for the specific components plot (default: "n_factor_specific")}
}
\description{
Plot posterior means of the loading matrices, including the response loadings in the supervised case.
Rotational alignment must be performed in advance.
To this end, make sure to provide in input the output of \code{\link{multiviewMatchAlign}}.
}
