% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jafar_plots.R
\name{plot_n_factors}
\alias{plot_n_factors}
\title{Plot MCMC samples of the inferred number of factors}
\usage{
plot_n_factors(
  risMCMC,
  out_path = "~/Desktop",
  out_name_shared = "n_factors_shared",
  out_name_specific = "n_factor_specific"
)
}
\arguments{
\item{risMCMC}{Posterior samples, output of \code{gibbs_jafar} or \code{gibbs_jfr}}

\item{out_path}{Output path where the generated plot will be saved}

\item{out_name_shared}{Output file name for the shared component plot (default: "n_factors_shared")}

\item{out_name_specific}{Output file name for the specific components plot (default: "n_factor_specific")}
}
\description{
Plot MCMC samples of the inferred number of factors
}
