% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jafar_plots.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Plot response predictions against true values}
\usage{
plot_predictions(
  yPred,
  yTrue,
  risMCMC,
  out_path = "~/Desktop/",
  out_name = "predictions"
)
}
\arguments{
\item{yPred}{Response predictions, output of \code{jafar_pred_y} or \code{jafar_coeff_y}}

\item{yTrue}{True values of the responses}

\item{risMCMC}{Posterior samples, output of \code{gibbs_jafar} or \code{gibbs_jfr}}

\item{out_path}{Output path where the generated plot will be saved (default: "~/Desktop/")}

\item{out_name}{Output file name (default: "predictions")}
}
\description{
Plot response predictions against true values
}
