% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jafar_preprocessing.R
\name{preprocess_X}
\alias{preprocess_X}
\title{Multi-view data preprocessing}
\usage{
preprocess_X(X_m, X_m_test = NULL, copula = FALSE)
}
\arguments{
\item{X_m}{Multi-view data for the training set.
List of length \code{M}; m-th element: matrix \code{n x p_m[m]}.}

\item{X_m_test}{Multi-view data for the test set.
List of length \code{M}; m-th element: matrix \code{nTest x p_m[m]}.
(optional, default: \code{NULL})}

\item{copula}{Apply cdf transformation (logical, default: \code{FALSE})}
}
\value{
List of pre-processed features and rescaling factors
}
\description{
Center and rescale, and apply feature-wise cdf transform (optional).
Multi-view data for out-of-sample observations can optionally be provided as input.
If so, the corresponding features are rescaled coherently with the training set.
}
\note{
The output list contains:
\itemize{
\item{\code{preprocess_X_m}: List of the rescaling transformations applied to the response,
as returned by \code{preProcess} from the \code{caret} package.
This object can be used to back-transform the data to the original scale.}
\item{\code{X_m}: Rescaled multi-view features for the training set.}
\item{\code{X_m_test}: Rescaled multi-view features for the test set (if provided).}
}
}
