% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jafar_preprocessing.R
\name{preprocess_y}
\alias{preprocess_y}
\title{Response preprocessing}
\usage{
preprocess_y(yTrain, yTest = NULL)
}
\arguments{
\item{yTrain}{Train set responses}

\item{yTest}{Test set responses (optional, default, \code{NULL})}
}
\value{
List of pre-processed responses and rescaling transformation
}
\description{
Center and rescale.
Response for out-of-sample observations can optionally be provided as input.
If so, the corresponding values are rescaled coherently with the training set.
}
\note{
The output list contains:
\itemize{
\item{\code{preprocess_y}: The rescaling transformation applied to the response,
as returned by \code{preProcess} from the \code{caret} package.
This object can be used to back-transform the data to the original scale.}
\item{\code{yTrain}: Rescaled responses for the training set.}
\item{\code{yTest}: Rescaled responses for the test set (if provided).}
}
}
