% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jafar_preprocessing.R
\name{preprocess_y}
\alias{preprocess_y}
\title{response preprocess: center & rescale}
\usage{
preprocess_y(yTrain, yTest = NULL)
}
\arguments{
\item{yTrain}{Train set responses}

\item{yTest}{Test set responses}
}
\value{
List of preprocessed responses and rescaling factors
}
\description{
response preprocess: center & rescale
}
